build: Write out stub passwd/group files so that getpwuid() works
Some bits of (broken) software attempt to look up the build user.  For
example, ibus presently runs dbus-launch to then invoke gsettings so
it can generate a text file.  dbus errors out when it's not able to
find its uid, of course.

We can help these bits of software along by generating an /etc/passwd
in the build environment that "works".

Signed-off-by: Pier Luigi Fiorini <pierluigi.fiorini@gmail.com>