build: Obey to forced build system.
Some components, like libpng, has multiple build systems but it's
not granted that the first one to be detected is the right one.
For example with libpng, CMake is the first one detected but we
don't have cmake yet and the build fails.

With this patch we can force a specific build system name, simply
adding the build-system metadata.