resolve: Ignore disabled components
Sometimes we want to temporarily disable some components but we don't
want to remove them from the manifest and since we can't comment the
JSON code (otherwise the parsing would fail) we need a mechanism
to the resolve that some components are gone.

This patch implements such mechanism, disabled components will be
removed from the source snapshot as they were removed from the
manifest.