snapshot: Filter out disabled components
Some components may not temporarily build, we usually pull a working
version but sometimes there's no working version at all.
That's the case for a few KF5 libraries. We want them in the manifest,
but we they are simply not yet ready.

We would like to comment such components out, but the JSON format
doesn't support comments, so we use the disabled key and exclude
this components from resolve so that they won't be even fetched.

This is a much better implementation than the one we had before,
which was removed with the recent overhaul.